Update WhiteLineFollowing
If motors are slowed to 100 the cornering doesnt work very effectively, with the speed increased the delays were reduced as well due to the acceleration being quicker. Also the overall changes to where the motor speed is defined is changed to allow for customization later. The addition for the testing so that every single combination of the three sensors is available to test they are correctly reading white from black.