Create OAWithSetDistances.ino
Works very well, runs fairly straight and completes the same turns every time. The values for these need to be tested, so the distance for 1 set of forward and back, then the angle of which right and left complete within a set period.  Once that is done the location can be calculated.