Update OAWithBatteryMonitoring.ino
Altered the battery monitoring limit as well, allowing the robot to function all the way. This value will need to be perfected based on its distance away from the docking station. But it runs straight, measures the battery. Just need to sort the distance values. I.e. measure the turn angle when it detects objects.